feat: EAGER, LAZY 전략에 따른 분기 처리
