+++ src/main/java/persistence/entity/loader/DefaultEntityLoader.java
@@ -1,7 +1,12 @@
 package persistence.entity.loader;
 
+import jakarta.persistence.FetchType;
+import java.lang.reflect.Field;
 import java.util.List;
 import jdbc.JdbcTemplate;
+import persistence.entity.loader.proxy.CollectionEntityProxy;
+import persistence.meta.ColumnMeta;
+import persistence.meta.RelationMeta;
 import persistence.meta.SchemaMeta;
 import persistence.sql.dml.query.WhereCondition;
 import persistence.sql.dml.query.WhereOperator;
@@ -31,9 +36,35 @@ public class DefaultEntityLoader implements EntityLoader {
             return instance;
         }
 
-        return collectionLoader.loadCollection(clazz, instance);
+        List<ColumnMeta> columnMetasWithRelation = schemaMeta.columnMetasHasRelation();
+        for (ColumnMeta columnMeta : columnMetasWithRelation) {
+            RelationMeta relationMeta = columnMeta.relationMeta();
+            if (FetchType.EAGER == relationMeta.fetchType()) {
+                List<?> children = eagerLoad(columnMeta, id);
+                mapChildrenField(instance, columnMeta, children);
+            }
+
+            if (FetchType.LAZY == relationMeta.fetchType()) {
+                List<?> children = CollectionEntityProxy.createProxy(columnMeta, id, collectionLoader);
+                mapChildrenField(instance, columnMeta, children);
+            }
+        }
+
+        return clazz.cast(instance);
     }
 
+    private List<?> eagerLoad(ColumnMeta columnMeta, Object id) {
+        return collectionLoader.load(columnMeta.relationMeta().joinColumnType(), columnMeta.field(), id);
+    }
 
+    private void mapChildrenField(Object instance, ColumnMeta columnMeta, List<?> children) {
+        try {
+            Field field = columnMeta.field();
+            field.setAccessible(true);
+            field.set(instance, children);
+        } catch (IllegalAccessException e) {
+            e.printStackTrace();
+        }
+    }
 
 }
