+++ src/main/java/persistence/entity/loader/proxy/CollectionEntityProxy.java
@@ -7,11 +7,11 @@ import java.util.List;
 import persistence.entity.loader.EntityCollectionLoader;
 import persistence.meta.ColumnMeta;
 
-public class CollectionEntityProxy implements InvocationHandler {
+public class CollectionEntityProxy<T> implements InvocationHandler {
 
     private final ColumnMeta columnMeta;
     private final Object id;
-    private List<?> target;
+    private List<T> target;
     private final EntityCollectionLoader entityLoader;
 
     public CollectionEntityProxy(ColumnMeta columnMeta, Object id, EntityCollectionLoader entityLoader) {
@@ -23,13 +23,13 @@ public class CollectionEntityProxy implements InvocationHandler {
     @Override
     public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
         if (targetIsNull()) {
-            target = entityLoader.load(columnMeta.relationMeta().joinColumnType(), columnMeta.field(), id);
+            target = (List<T>) entityLoader.load(columnMeta.relationMeta().joinColumnType(), columnMeta.field(), id);
         }
         return method.invoke(target, args);
     }
 
-    public static List<?> createProxy(ColumnMeta columnMeta, Object id, EntityCollectionLoader entityLoader) {
-        return (List<?>) Proxy.newProxyInstance(
+    public static <T> List<T> createProxy(ColumnMeta columnMeta, Object id, EntityCollectionLoader entityLoader) {
+        return (List<T>) Proxy.newProxyInstance(
                         columnMeta.field().getType().getClassLoader(),
                         new Class[]{List.class},
                         new CollectionEntityProxy(columnMeta, id, entityLoader));
